test: reorganise test import to prevent circular dependency
