+++ src/ares/behaviors/combat/individual/use_ability.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Union
 
 from sc2.ids.ability_id import AbilityId
 from sc2.position import Point2
@@ -41,7 +41,7 @@ class UseAbility(CombatIndividualBehavior):
 
     ability: AbilityId
     unit: Unit
-    target: Optional[Point2, Unit] = None
+    target: Optional[Union[Point2, Unit]] = None
 
     def execute(
         self, ai: "AresBot", config: dict, mediator: ManagerMediator, **kwargs
